remove extraneous stuff and reorg
