+++ docker-compose.yml
@@ -3,7 +3,7 @@ version: '3'
 services:
 
   api:
-    container_name: api # Name can be anything
+    container_name: api
     restart: always
     build: ./api
     ports:
