+++ nginx/Dockerfile
@@ -1,4 +1,3 @@
-# Docker file for 
 FROM nginx:1.15.2
 
 RUN rm /etc/nginx/nginx.conf
