+++ site/templates/results.html
@@ -1,0 +1,37 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset='utf-8' />
+    <meta http-equiv='content-type' content='text/html; charset=utf-8' />
+
+    <title>Contention Classifier</title>
+
+   
+
+  </head>
+  <body>
+      <center>
+          <br>
+          <h1>Contention Classifier</h1>
+          <br>
+               
+          
+      
+          <p><h4><font color="black">Contention:</font> <font color="blue">{{results["text"]}} </font></h4><p> 
+          <p><h4><font color="black">Classification:</font> <font color="blue">{{results["prediction"]["classification"]}} </font></h4><p> 
+          <p><h4><font color="black">Classification Code:</font> <font color="blue">{{results["prediction"]["code"]}} </font></h4><p> 
+          <p><h4><font color="black">Confidence in Prediction:</font> <font color="blue">{{results["prediction"]["probability"]}} </font></h4><p> 
+
+           
+          <br>
+          <a href="/index"><font color="black">Go back</font></a>
+          <br>
+          <br>
+          <br>
+          <br>
+          <br>
+          <br>
+      </center>
+  </body>
+
+</html>
