--- templates/index.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<meta charset="utf-8">
-
-
-<body>
-    <center><h1>Contention Classifier</h1></center>
-    
-
-
-    
-    
-    <center>
-     <h4>Enter a contention:</h4>
-    <form id='input' method='post' action='index'>	
-	  			<p>
-					<input type='text' name='claim_text' />
-        </p>
-
-	    			<input type="submit" value="Submit">
-	  		
-				</form>
-    </center>
-    
-</body>
